requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
sqlalchemy==2.0.35
pydantic==2.9.2
python-dotenv==1.0.1
uuid==1.30
agora-token-builder
slowapi==0.1.9
httpx>=0.27.0
pytest==8.3.5